Move pop up function to index.js and generated main.js using webpack
